feat: sdl 3 and additions (#14)
* feat: sdl3

* feat: random module

* feat: random module

* fix: ttf init

* feat: add scene manager

* feat: add fps

* fix: template scene

* feat: emscripten support

* fix: mouse movement

* chore: update cmake

* chore: update cmake

* fix: add vendor option

* fix: add vendor option

* fix: scale coordinates

* chore: fix order

* fix: backward compat

* feat: add geom primitives

* feat: float mouse

* feat: quad ops

* feat: add vsync

* fix: letterboxing

* feat: add game module

* feat: register object

* fix: scene logic

* fix: cleanup game objects

* chore: remove entity id

* fix: sonar

+++ include/asw/modules/assets.h
@@ -1,12 +1,11 @@
-/**
- * @file assets.h
- * @author Allan Legemaate (alegemaate@gmail.com)
- * @brief Asset routines for the ASW library
- * @date 2023-09-20
- *
- * @copyright Copyright (c) 2023
- *
- */
+/// @file assets.h
+/// @author Allan Legemaate (alegemaate@gmail.com)
+/// @brief Asset routines for the ASW library
+/// @date 2023-09-20
+///
+/// @copyright Copyright (c) 2023
+///
+
 #ifndef ASW_ASSETS_H
 #define ASW_ASSETS_H
 
@@ -16,42 +15,39 @@
 #include "./util.h"
 
 namespace asw::assets {
-  /**
-   * @brief Loads a texture from a file. Formats supported are PNG, ICO, CUR,
-   * BMP, GIF, JPG, LBM, PCX, PNM, TIF, XCF, XPM, XV, and WEBP. This will abort
-   * if the file is not found.
-   *
-   * @param filename The path to the file.
-   * @return The loaded texture.
-   */
+
+  /// @brief Loads a texture from a file. Formats supported are PNG, ICO, CUR,
+  /// BMP, GIF, JPG, LBM, PCX, PNM, TIF, XCF, XPM, XV, and WEBP. This will abort
+  /// if the file is not found.
+  ///
+  /// @param filename The path to the file.
+  /// @return The loaded texture.
+  ///
   asw::Texture loadTexture(const std::string& filename);
 
-  /**
-   * @brief Loads a TTF font from a file. This will abort if the file is not
-   * found.
-   *
-   * @param filename The path to the file.
-   * @param size The size of the font.
-   * @return The loaded font.
-   */
-  asw::Font loadFont(const std::string& filename, int size);
+  /// @brief Loads a TTF font from a file. This will abort if the file is not
+  /// found.
+  ///
+  /// @param filename The path to the file.
+  /// @param size The size of the font.
+  /// @return The loaded font.
+  ///
+  asw::Font loadFont(const std::string& filename, float size);
 
-  /**
-   * @brief Loads a sample from a file. Formats supported are WAV, AIFF, RIFF,
-   * OGG and VOC. This will abort if the file is not found.
-   *
-   * @param filename
-   * @return The loaded sample.
-   */
+  /// @brief Loads a sample from a file. Formats supported are WAV, AIFF, RIFF,
+  /// OGG and VOC. This will abort if the file is not found.
+  ///
+  /// @param filename
+  /// @return The loaded sample.
+  ///
   asw::Sample loadSample(const std::string& filename);
 
-  /**
-   * @brief Create a Texture given the specified dimensions.
-   *
-   * @param w The width of the texture.
-   * @param h The height of the texture.
-   * @return The created texture.
-   */
+  /// @brief Create a Texture given the specified dimensions.
+  ///
+  /// @param w The width of the texture.
+  /// @param h The height of the texture.
+  /// @return The created texture.
+  ///
   asw::Texture createTexture(int w, int h);
 }  // namespace asw::assets
